+++ contrib/prcycoin-qt.pro
@@ -149,6 +149,7 @@ HEADERS += src/activemasternode.h \
            src/wallet/wallet_ismine.h \
            src/wallet/walletdb.h \
            src/consensus/consensus.h \
+           src/consensus/validation.h \
            src/compat/sanity.h \
            src/config/prcycoin-config.h \
            src/crypto/common.h \
