+++ src/main.cpp
@@ -13,6 +13,7 @@
 #include "chainparams.h"
 #include "checkpoints.h"
 #include "checkqueue.h"
+#include "consensus/validation.h"
 #include "fs.h"
 #include "init.h"
 #include "kernel.h"
